Remove side panel borders and move buffer slider to stats
- Drop the vertical dividers on the moq-boy side panels so the game
  canvas blends with the controls/stats.
- Move the buffer/jitter slider from the Controls panel into the
  StatsPanel, where it sits next to the latency breakdown.
- Bump @moq/boy to 0.2.4.
